Uint32 Seed
Justification: The seed internally is changed between different number types with wildly different max values. This PR normalizes all of them to UInt32.

Earlier @pcuenca commented that the reason for the lower precision use of Double was to support the CompactSlider component and make it easier for a user to select a number.

In this submission the CompactSlider component has been swapped for a TextField.

[Feature extracted from larger PR #59 https://github.com/huggingface/swift-coreml-diffusers/pull/59]

+++ Diffusion-macOS/StatusView.swift
@@ -87,10 +87,10 @@ struct StatusView: View {
                 let intervalString = String(format: "Time: %.1fs", interval ?? 0)
                 Text(intervalString)
                 Spacer()
-                if generation.seed != Double(lastSeed) {
+                if generation.seed != UInt32(lastSeed) {
                     Text("Seed: \(lastSeed)")
                     Button("Set") {
-                        generation.seed = Double(lastSeed)
+                        generation.seed = UInt32(lastSeed)
                     }
                 }
             }.frame(maxHeight: 25)
